# openapi.yaml
openapi: 3.1.0
info:
  title: NeroCD API
  version: 0.1.0
  description: |
    Public REST API for the NeroCD automation control plane.

    All public routes are versioned under `/api/v1`. List responses use a stable
    `{ "items": [...] }` envelope. Error responses use
    `{ "code": "stable_code", "error": "message" }`.

    List endpoints return a stable `{ "items": [...] }` envelope plus
    pagination metadata: `limit`, `offset`, `count`, and `total`. Clients can
    request `limit` and `offset` query parameters. Filtering is documented per
    endpoint. Mutations consume JSON request bodies and return the created
    resource or validation errors.
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
security:
  - bearerAuth: []
  - browserSession: []
paths:
...
  /api/v1/run-log-retention:
    get:
      summary: Get the global manual run-log retention policy and aggregate preview
      operationId: getRunLogRetentionStatus
      responses:
        "200": {description: Current policy and aggregate-only preview, content: {application/json: {schema: {$ref: "#/components/schemas/RunLogRetentionStatus"}}}}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
    put:
      summary: Update the global manual run-log retention policy
      operationId: updateRunLogRetentionPolicy
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/RunLogRetentionPolicyInput"}}}
      responses:
        "200": {description: Updated policy and aggregate-only preview, content: {application/json: {schema: {$ref: "#/components/schemas/RunLogRetentionStatus"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /api/v1/run-log-retention/execute:
    post:
      summary: Execute one bounded, idempotent manual run-log retention batch
      operationId: executeRunLogRetention
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/RunLogRetentionExecuteInput"}}}
      responses:
        "200": {description: Immutable execution receipt with aggregate-only result, content: {application/json: {schema: {$ref: "#/components/schemas/RunLogRetentionExecution"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /api/v1/projects:
    get:
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      summary: List projects
      operationId: listProjects
      responses:
        "200":
          description: Project list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    post:
      summary: Create project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update project
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: string
                - $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/projects/archive:
    post:
      summary: Archive project
      operationId: archiveProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IDRequest"
      responses:
        "200":
          description: Archived project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Error"
...
  /api/v1/repositories:
    get:
      summary: List source repositories
      operationId: listRepositories
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Repository list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      summary: Create source repository
      operationId: createRepository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryInput"
      responses:
        "201":
          description: Created repository.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/repositories/{id}/policy:
    put:
      summary: Configure a legacy repository source policy
      operationId: configureRepositoryPolicy
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
         application/json:
           schema:
             type: object
             required: [project_id, configuration_id, policy]
             additionalProperties: false
             properties:
               project_id: {type: string}
               configuration_id: {type: string, pattern: '^cfg_[A-Za-z0-9_-]{8,128}$'}
               policy: {$ref: "#/components/schemas/RepositoryPolicy"}
      responses:
        "200": {description: Configured repository, content: {application/json: {schema: {$ref: "#/components/schemas/Repository"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/project-members:
    get:
      summary: List project members
      operationId: listProjectMembers
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Project member list. When `project_id` is supplied, only members for that project are returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMemberListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      summary: Grant or update project access
      operationId: upsertProjectMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectMemberInput"
      responses:
        "200":
          description: Project member access updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/v1/access-keys:
    get:
      summary: List project access key metadata
      operationId: listAccessKeys
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Access key metadata list. Secret values are never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessKeyListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      summary: Create project access key metadata
      operationId: createAccessKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessKeyInput"
      responses:
        "201":
          description: Created access key metadata. Secret values are not accepted or returned by this endpoint.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessKey"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/inventories:
    get:
      summary: List project inventory metadata
      operationId: listInventories
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Inventory metadata list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      summary: Create project inventory metadata
      operationId: createInventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryInput"
      responses:
        "201":
          description: Created inventory metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: NeroCD session token
      description: CLI session and API token credential.
    browserSession:
      type: apiKey
      in: cookie
      name: nerocd_session
      description: "Browser-only local session cookie. Unsafe cookie-authenticated requests require X-NeroCD-CSRF: 1."
  parameters:
    ProjectID:
      name: project_id
      in: query
      required: false
      description: Optional project identifier filter.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return (0-100). Omit to use the bounded default.
      schema:
        type: integer
        minimum: 0
        maximum: 100
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip from the filtered result.
      schema:
        type: integer
        minimum: 0
        maximum: 100000
  responses:
    Error:
      description: Stable error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    RunLogRetentionPolicy:
      type: object
      additionalProperties: false
      required: [enabled, keep_days, batch_size, version, updated_at]
      properties:
        enabled: {type: boolean}
        keep_days: {type: integer, minimum: 1, maximum: 3650}
        batch_size: {type: integer, minimum: 1, maximum: 10000}
        version: {type: integer, minimum: 1}
        updated_at: {type: string, format: date-time}
    RunLogRetentionPolicyInput:
      type: object
      additionalProperties: false
      required: [enabled, keep_days, batch_size]
      properties:
        enabled: {type: boolean}
        keep_days: {type: integer, minimum: 1, maximum: 3650}
        batch_size: {type: integer, minimum: 1, maximum: 10000}
    RunLogRetentionPreview:
      type: object
      additionalProperties: false
      required: [cutoff, eligible_logs, eligible_bytes]
      properties:
        cutoff: {type: string, format: date-time}
        eligible_logs: {type: integer, minimum: 0, format: int64}
        eligible_bytes: {type: integer, minimum: 0, format: int64}
    RunLogRetentionStatus:
      type: object
      additionalProperties: false
      required: [policy, preview]
      properties:
        policy: {$ref: "#/components/schemas/RunLogRetentionPolicy"}
        preview: {$ref: "#/components/schemas/RunLogRetentionPreview"}
    RunLogRetentionExecuteInput:
      type: object
      additionalProperties: false
      required: [policy_version]
      properties:
        policy_version: {type: integer, minimum: 1}
    RunLogRetentionExecution:
      type: object
      additionalProperties: false
      required: [request_id, policy, preview, deleted, deleted_bytes]
      properties:
        request_id: {type: string}
        policy: {$ref: "#/components/schemas/RunLogRetentionPolicy"}
        preview: {$ref: "#/components/schemas/RunLogRetentionPreview"}
        deleted: {type: integer, minimum: 0, format: int64}
        deleted_bytes: {type: integer, minimum: 0, format: int64}
    ErrorResponse:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          enum: [unauthenticated, forbidden, not_found, bad_request, internal_error]
        error:
          type: string
...
    Project:
      type: object
      required: [id, name, description, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    IDRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
...
    ProjectListResponse:
      type: object
      required: [items, limit, offset, count, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        count:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
    ProjectMember:
      type: object
      required: [id, project_id, user_id, email, name, role, created_at, updated_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        user_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [owner, maintainer, viewer]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectMemberInput:
      type: object
      required: [project_id, email, role]
      properties:
        project_id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [owner, maintainer, viewer]
    ProjectMemberListResponse:
      type: object
      required: [items, limit, offset, count, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        count:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    AccessKey:
      type: object
      required: [id, project_id, name, kind, fingerprint, created_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [ssh, password, token]
        fingerprint:
          type: string
        created_at:
          type: string
          format: date-time
    AccessKeyInput:
      type: object
      required: [project_id, name, kind, fingerprint]
      properties:
        project_id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [ssh, password, token]
        fingerprint:
          type: string
    AccessKeyListResponse:
      type: object
      required: [items, limit, offset, count, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AccessKey"
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        count:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
    Inventory:
      type: object
      required: [id, project_id, name, kind, source, created_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [static, dynamic]
        source:
          type: string
        created_at:
          type: string
          format: date-time
    InventoryInput:
      type: object
      required: [project_id, name, kind, source]
      properties:
        project_id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [static, dynamic]
        source:
          type: string
    InventoryListResponse:
      type: object
      required: [items, limit, offset, count, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Inventory"
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        count:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    Repository:
      type: object
      required: [id, project_id, name, url, provider, default_ref, policy, created_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        name:
          type: string
        url:
          type: string
        provider:
          type: string
        default_ref:
          type: string
        policy: {$ref: "#/components/schemas/RepositoryPolicy"}
        created_at:
          type: string
          format: date-time
    RepositoryInput:
      type: object
      required: [project_id, name, url]
      properties:
        project_id:
          type: string
        name:
          type: string
        url:
          type: string
        provider:
          type: string
        default_ref:
          type: string
        policy: {$ref: "#/components/schemas/RepositoryPolicy"}
    RepositoryPolicy:
      discriminator: {propertyName: state}
      oneOf:
        - type: object
          required: [version, state]
          properties: {version: {type: integer, enum: [1]}, state: {type: string, enum: [legacy_unverified]}}
          additionalProperties: false
        - type: object
          required: [version, state, mode, allowed_schemes, allowed_hosts]
          additionalProperties: false
          properties:
            version: {type: integer, enum: [1]}
            state: {type: string, enum: [configured]}
            mode: {type: string, enum: [public, internal]}
            allowed_schemes: {type: array, minItems: 1, maxItems: 2, items: {type: string, enum: [https, ssh, http]}}
            allowed_hosts: {type: array, minItems: 1, maxItems: 32, items: {type: string}}
            allowed_cidrs: {type: array, maxItems: 32, items: {type: string}}
            redirect_hosts: {type: array, maxItems: 32, items: {type: string}}
            ssh_host_fingerprints: {type: array, maxItems: 32, items: {type: string}}
            credential_reference_id: {type: string, pattern: '^cred_[A-Za-z0-9_-]{8,128}$'}
            allow_internal: {type: boolean}
    RepositoryListResponse:
      type: object
      required: [items, limit, offset, count, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Repository"
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        count:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...